% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{autofix}
\alias{autofix}
\title{Autofix}
\usage{
autofix(W, copy = F)
}
\arguments{
\item{W}{: an R matrix - weighted connectivity matrix}

\item{copy}{: boolean}
}
\value{
W : an R matrix - autofixed weighted connectivity matrix
}
\description{
Fix a bunch of common problems. 
More specifically, remove Inf and NaN, ensure 
exact binariness and symmetry (i.e. remove floating point
instability), and zero diagonal.
}

