% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{binarize}
\alias{binarize}
\title{Binarize}
\usage{
binarize(W, copy = F)
}
\arguments{
\item{W}{: an R matrix - weighted connectivity matrix}

\item{copy}{: boolean}
}
\value{
W : an R matrix - binary weighted connectivity matrix
}
\description{
Binarizes an input weighted connection matrix,
i.e. any value > 0 gets set to 1
}

