% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{breadth}
\alias{breadth}
\title{Breadth-First Search}
\usage{
breadth(CIJ, src)
}
\arguments{
\item{CIJ}{: a matrix - binary (un)directed connection matrix}

\item{src}{: an integer - the source vertex index}
}
\value{
DB : a list - DB has two elements: 'distance' is a vector of
distances between the source (src) and the ith vertex (0 for src),
and 'branch' is a vector of vertices that precede 'i' in the BFS (-1 for src)
}
\description{
Implementation of Breadth-First Search in R (yikes!)
}

