% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{breadthdist}
\alias{breadthdist}
\title{Binary Reachibility Matrix}
\usage{
breadthdist(CIJ)
}
\arguments{
\item{CIJ}{: a matrix - binary (un)directed connection matrix}
}
\value{
RD : a list - RD has two elements: 'R' is the binary
reachability matrix, and 'D' is the distance matrix.
}
\description{
The binary reachability matrix 'R' describes reachibility
between all pairs of nodes. An entry (u,v)=1 means that
there exists a path from node u to node v, and (u,v)=0
means there does not.
}
\details{
The distance matrix 'D' contains lengths of shortest paths
between all pairs of nodes. An entry (u,v) represents
the length of shortest path from node u to node v. The
average shortest path length is the characteristic
path length of the network.
}

