% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clustering.coef.wu}
\alias{clustering.coef.wu}
\title{Weighted Undirected Clustering Coefficient}
\usage{
clustering.coef.wu(W)
}
\arguments{
\item{W}{: a matrix - weighted undirected connection matrix}
}
\value{
C : a vector - clustering coefficient vector
}
\description{
The weighted Und. clustering coefficient is the
average "intensity" of triangles around a node
}

