% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get.components}
\alias{get.components}
\title{Connected Components of an Undirected Graph}
\usage{
get.components(A)
}
\arguments{
\item{A}{: a matrix - binary undirected adjacency matrix}
}
\value{
comp : a list - each list element is a graph component, and
each component has two elements: 'members' - the vector of nodes belonging
to that component, and 'size' - the number of nodes in that component
}
\description{
Returns the components of an undirected graph specified by the 
binary and undirected adjacency matrix 'A'. Components and their
constituent nodes are assigned the same index and stored in the
vector 'comp'. The vector 'comp.sizes' contains the number of
nodes belonging to each component. Furthermore, 'members' and
'size' are combined into one list ' comp
}
\details{
Disconnected nodes will appear as components with a component
size of 1.

The identity of each component (i.e. its numerical value in the 
result) is not guaranteed to be identical to the value return in
other versions of BCT, although the component topology is.

In the Python-BCT, networkx is used to do the computation efficiently.
In the Matlab-BCT, the Dulmage-Mendelsohn decomposition is computed.

Note: Python version takes 1.8s w/out networkx and 54ms w/ networkx w/ sample.data
}

