% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{invert}
\alias{invert}
\title{Invert}
\usage{
invert(W, copy = F)
}
\arguments{
\item{W}{: an R matrix - weighted connectivity matrix}

\item{copy}{: boolean}
}
\value{
W : an R matrix - inverted weighted connectivity matrix
}
\description{
Inverts elementwise the weights in an input connection matrix,
i.e. changes W from a matrix of internode strengths to 
a matrix of internode distances
}

