% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(W, copy = F)
}
\arguments{
\item{W}{: an R matrix - weighted connectivity matrix}

\item{copy}{: boolean}
}
\value{
W : an R matrix - normalized weighted connectivity matrix
}
\description{
Normalizes an input weighted connection matrix,
i.e. divides each value by the abs. max. value in matrix
}

