% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability_test_retest}
\alias{reliability_test_retest}
\title{reliability study}
\usage{
reliability_test_retest()
}
\value{

}
\description{
A vector of measurements in test,
}
\references{
https://en.wikipedia.org/wiki/Repeatability
}

