% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICE_proportional.R
\name{SICE_proportional}
\alias{SICE_proportional}
\title{Brain Network Construction via SICE}
\usage{
SICE_proportional(W, p, rho_init = 1, rho_min = 0, rho_max = 5,
  errRange = 10)
}
\description{
Brain Network Construction via SICE
}

