% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{threshold.absolute}
\alias{threshold.absolute}
\title{Threshold Absolute}
\usage{
threshold.absolute(W, thr, copy = F)
}
\arguments{
\item{W}{: an R Matrix - Weighted Connectivity Matrix}

\item{thr}{: a float - Absolute Weight Threshold}

\item{copy}{: a boolean - Whether to modify in place or not}
}
\value{
W : Threshold Connectivity Matrix
}
\description{
This function thresholds the connectivity matrix by absolute weight
magnitude. All weights below the given threshold, and all weights
on the main diagonal (self-self connections) are set to 0.
}

