% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_proportional.R
\name{threshold_proportional}
\alias{threshold_proportional}
\title{Brain Network Construction via threshold}
\usage{
threshold_proportional(W, p)
}
\arguments{
\item{W}{covariance matrix.}

\item{p}{links percentage.}
}
\value{
thresholded connectivity matrix.
}
\description{
Construct brain connectivity network
w/ specified links percentage via threshold
}
\examples{
 library(Matrix)
 x <- runif(20, 1, 5)
 mat <- matrix(x, nrow = 4)
 cov <-  cov(mat)
 threshold_proportional(cov, 0.3)

}

