% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{transitivity.wu}
\alias{transitivity.wu}
\title{Weighted Undirected Transitivity}
\usage{
transitivity.wu(W)
}
\arguments{
\item{W}{: a matrix - weighted undirected connection matrix}
}
\value{
t : an integer - transitivity scalar
}
\description{
Transitivity is the ratio of 'triangles to triplets' in the network.
(A classical version of the clustering coefficient).
}

