% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{weight.conversion}
\alias{weight.conversion}
\title{Weight Conversion}
\usage{
weight.conversion(W, wcm, copy = F)
}
\arguments{
\item{W}{: an R Matrix - Weighted Connectivity Matrix}

\item{wcm}{: a string - Weight Conversion Command}
}
\value{
W : an R Matrix - Connectivity Matrix with 
specified changes
}
\description{
This function may either binarize an input weighted connection matrix,
normalize an input weighted connection matrix or convert an input
weighted connection matrix to a weighted connection-length matrix.
}
\details{
Binarization converts all present connection weights to 1.

Normalization scales all weight magnitudes to the range [0,1] and
should be done prior to computing some weighted measures, such as the
weighted clustering coefficient.

Conversion of connection weights to connection lengths is needed
prior to computation of weighted distance-based measures, such as
distance and betweenness centrality. In a weighted connection network,
higher weights are naturally interpreted as shorter lengths. The
connection-lengths matrix here is defined as the inverse of the
connection-weights matrix.

Options for 'wcm' param:
       'binarize' : binarize weights
       'normalize' : normalize weights
       'lengths' : convert weights to lengths (invert matrix)
}

